// src/components/page-entry/page-entry.interface.ts
interface MediaItem {
  thumbnail: string;
  source: string;
  type: 'image' | 'video';
}

export interface PageEntryProps {
  media: MediaItem[];
  headline: string;
  tagline: string;
  description: string;
  primaryButtonText: string;
  primaryButtonClick: () => void;
  secondaryButtonText: string;
  secondaryButtonClick: () => void;
}